commit validacoes

+++ br.com.bureau/src/main/java/br/com/bureau/service/base_a/PersonService.java
@@ -6,15 +6,49 @@ import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import br.com.bureau.exception.ConflictedException;
 import br.com.bureau.exception.NotFoundException;
+import br.com.bureau.models.base_a.Address;
+import br.com.bureau.models.base_a.Debts;
 import br.com.bureau.models.base_a.Person;
+import br.com.bureau.models.base_b.Assets;
+import br.com.bureau.models.base_b.Income;
+import br.com.bureau.models.base_c.LastBuy;
+import br.com.bureau.models.base_c.LastSearch;
+import br.com.bureau.models.base_c.Movement;
 import br.com.bureau.repository.base_a.PersonRepository;
+import br.com.bureau.service.base_b.AssetsService;
+import br.com.bureau.service.base_b.IncomeService;
+import br.com.bureau.service.base_c.LastBuyService;
+import br.com.bureau.service.base_c.LastSearchService;
+import br.com.bureau.service.base_c.MovementService;
 
 @Service
 public class PersonService {
 
 	@Autowired
 	private PersonRepository personRepository;
+	
+	@Autowired
+	private AddressService addressService;
+	
+	@Autowired
+	private DebtsService debtsService;
+	
+	@Autowired
+	private AssetsService assetsService;
+	
+	@Autowired
+	private IncomeService incomeService;
+	
+	@Autowired
+	private LastBuyService buyService;
+	
+	@Autowired
+	private LastSearchService lastSearchService;
+	
+	@Autowired
+	private MovementService movementService;
 	
 	public Person create(Person person) {
 		return personRepository.save(person);
@@ -40,6 +74,34 @@ public class PersonService {
 	
 	public void delete(Long id) {
 		Person person  = findById(id);
+		List<Address> addresses = addressService.getAddressByPerson(id);
+		if (addresses == null || !addresses.isEmpty()) {
+			throw new ConflictedException("A pessoa não pode ser apagada porque existem endereços cadastrados");
+		}
+		List<Debts> debts = debtsService.getDebtsByPerson(id);
+		if (debts == null || !debts.isEmpty()) {
+			throw new ConflictedException("A pessoa não pode ser apagada porque existem débitos cadastrados");
+		}
+		List<Assets> assets = assetsService.getAssetsByPerson(id);
+		if (assets == null || !assets.isEmpty()) {
+			throw new ConflictedException("A pessoa não pode ser apagada porque existem bens cadastrados");
+		}
+		List<Income> income = incomeService.getIncomeByPerson(id);
+		if (income == null || !income.isEmpty()) {
+			throw new ConflictedException("A pessoa não pode ser apagada porque existem rendas cadastradas");
+		}
+		List<LastBuy> lastBuy = buyService.getLastBuyByPerson(id);
+		if (lastBuy == null || !lastBuy.isEmpty()) {
+			throw new ConflictedException("A pessoa não pode ser apagada porque existem compras cadastradas");
+		}
+		List<LastSearch> lastSearch = lastSearchService.getLastSearchByPerson(id);
+		if (lastSearch == null || !lastSearch.isEmpty()) {
+			throw new ConflictedException("A pessoa não pode ser apagada porque existem pesquisas cadastrados");
+		}
+		List<Movement> movement = movementService.getMovementByPerson(id);
+		if (movement == null || !movement.isEmpty()) {
+			throw new ConflictedException("A pessoa não pode ser apagada porque existem movimentações cadastrados");
+		}
 		this.personRepository.delete(person);
 	}
